requirements.txt
langhcahin
openai
tiktoken
tqdm
pinecone-client
streamlit
streamlit_chat